+++ crates/cuid1/benches/cuid.rs
@@ -1,3 +1,4 @@
+#![allow(deprecated)]
 #[macro_use]
 extern crate criterion;
 extern crate cuid;