+++ crates/cuid1/tests/cuid1/main.rs
@@ -1,0 +1,6 @@
+//! Test entrypoint for (deprecated) cuid() function
+
+#![allow(deprecated)]
+
+mod collisions;
+mod length;